pyproject.toml
[tool.poetry]
name = "ea-sim"
version = "0.1.1"
description = "Similarity model"
authors = [
]
readme = "README.md"
packages = [
    {include = "ea"}
]
exclude = [
    "ea/sim/dev"
]
[tool.poetry.dependencies]
python = "3.10.*"
numpy = "^1.24.2"
pandas = "<2.0.0"
tqdm = "^4.65.0"
matplotlib = "^3.7.1"
scikit-learn = "^1.2.2"
hyperopt = "^0.2.7"
faiss-cpu = "^1.7.3"
loguru = "^0.7.0"
cython = "^0.29.34"
scipy = "^1.10.1"
ea-common = { "path" = "../common", develop = true }
pytest = "^7.4.1"
lightning = "2.1.1"
tokenizers = "^0.15.0"
plotly = "^5.19.0"
pytorch-metric-learning = "^2.5.0"
transformers = "^4.39.3"

[tool.poetry.group.dev.dependencies]
mlflow = "^2.6.0"
jupyter = "^1.0.0"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
